Thu, Jan 12, 2023, 3:52 PM +00:00

+++ filter-less/helpers.c
@@ -35,9 +35,9 @@ void sepia(int height, int width, RGBTRIPLE image[height][width])
         for (int j = 0; j < width; j++)
         {
             // Using sepia's algorithm. Casting first values to an int so the programm doesn't run out of bits.
-            redsepia = round (((int) .393 * image[i][j].rgbtRed) + (.769 * image[i][j].rgbtGreen) + (.189 * image[i][j].rgbtBlue));
-            greensepia = round (((int) .349 * image[i][j].rgbtRed) + (.686 * image[i][j].rgbtGreen) + (.168 * image[i][j].rgbtBlue));
-            bluesepia = round (((int) .272 * image[i][j].rgbtRed) + (.534 * image[i][j].rgbtGreen) + (.131 * image[i][j].rgbtBlue));
+            redsepia = round (( .393 * image[i][j].rgbtRed) + (.769 * image[i][j].rgbtGreen) + (.189 * image[i][j].rgbtBlue));
+            greensepia = round (( .349 * image[i][j].rgbtRed) + (.686 * image[i][j].rgbtGreen) + (.168 * image[i][j].rgbtBlue));
+            bluesepia = round (( .272 * image[i][j].rgbtRed) + (.534 * image[i][j].rgbtGreen) + (.131 * image[i][j].rgbtBlue));
             // Assigning sepia values to the original pixel values
             image[i][j].rgbtRed = redsepia;
             image[i][j].rgbtGreen = greensepia;
